# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Insider One - Notification API",
        "description": "Toplu bildirim ve şablon yönetimi için REST API dokümantasyonu.",
        "contact": {
            "name": "Insider Dev Team",
            "email": "dev@insider.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/notification/{id}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "ID'ye göre bildirim getir",
                "operationId": "notification.findById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bildirim",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Bulunamadı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Alıcıya ait bildirimi güncelle",
                "description": "Belirtilen alıcıya ait bildirim kayıtlarını günceller.",
                "operationId": "notification.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bildirim ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "7bb61a62-1263-406a-b20d-081aff51ad5d"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "recipient": {
                                        "type": "string",
                                        "example": "905321234567"
                                    },
                                    "channel": {
                                        "type": "string",
                                        "example": "sms"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "sent"
                                    },
                                    "batch_id": {
                                        "type": "string",
                                        "example": "123e4567-e89b-12d3-a456-426614174000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bildirim güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasyon hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Sunucu hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Alıcıya ait bildirimi sil",
                "description": "Belirtilen alıcıya ait tüm bildirim kayıtlarını siler.",
                "operationId": "notification.delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Bildirim ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bildirim silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Sunucu hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/notification-template": {
            "post": {
                "tags": [
                    "Notification Templates"
                ],
                "summary": "Bildirim şablonu oluştur",
                "description": "Yeni bir bildirim şablonu oluşturur.",
                "operationId": "notificationTemplate.insert",
                "requestBody": {
                    "description": "Şablon bilgileri",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "order_shipped"
                                    },
                                    "channel": {
                                        "type": "string",
                                        "example": "sms"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Siparişiniz kargoya verildi."
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "order_shipped",
                                "channel": "sms",
                                "content": "Siparişiniz kargoya verildi.",
                                "status": "active"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Şablon başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasyon hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Sunucu hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/notification-template/{id}": {
            "get": {
                "tags": [
                    "Notification Templates"
                ],
                "summary": "ID ile şablon getir",
                "description": "UUID ile bir bildirim şablonunu getirir.",
                "operationId": "notificationTemplate.findById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Şablon UUID'si",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Şablon bulundu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Şablon bulunamadı veya sunucu hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Notification Templates"
                ],
                "summary": "Bildirim şablonunu güncelle",
                "description": "Belirtilen UUID'ye sahip şablonu günceller.",
                "operationId": "notificationTemplate.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Şablon UUID'si",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/NotificationTemplateItem"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Şablon güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasyon hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Sunucu hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notification Templates"
                ],
                "summary": "Bildirim şablonunu sil",
                "description": "Belirtilen UUID'ye sahip şablonu siler.",
                "operationId": "notificationTemplate.delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Şablon UUID'si",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "550e8400-e29b-41d4-a716-446655440000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Şablon silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Sunucu hatası",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "NotificationTemplateItem": {
                "required": [
                    "name",
                    "channel",
                    "content"
                ],
                "properties": {
                    "name": {
                        "description": "Şablon adı",
                        "type": "string",
                        "example": "order_shipped"
                    },
                    "channel": {
                        "description": "Kanal",
                        "type": "string",
                        "enum": [
                            "sms",
                            "email",
                            "push"
                        ],
                        "example": "email"
                    },
                    "content": {
                        "description": "Şablon içeriği",
                        "type": "string",
                        "example": "Merhaba, siparişiniz kargoya verilmiştir."
                    },
                    "status": {
                        "description": "Şablon durumu",
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    }
                },
                "type": "object"
            },
            "BaseResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    },
                    "data": {
                        "nullable": true
                    },
                    "code": {
                        "type": "integer",
                        "example": 200
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "data": {
                        "example": null,
                        "nullable": true
                    },
                    "code": {
                        "type": "integer",
                        "example": 400
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}